Send Study UID while trying to share study, API call to get XRay results and cleaning up
