+++ extensions/cornerstone-dicom-sr/src/getSopClassHandlerModule.ts
@@ -172,7 +172,6 @@ async function _load(displaySet, servicesManager: AppTypes.ServicesManager, exte
     displaySet.referencedImages = [];
     displaySet.measurements = [];
   }
-  const l = _procedureReportExists(ContentSequence);
 
   const mappings = measurementService.getSourceMappings(
     CORNERSTONE_3D_TOOLS_SOURCE_NAME,
@@ -350,19 +349,7 @@ function getSopClassHandlerModule({ servicesManager, extensionManager }) {
     },
   ];
 }
-function _procedureReportExists(ProceduretReportContentSequence) {
-  const Procedure = ProceduretReportContentSequence.find(
-    item => item.ConceptNameCodeSequence.CodeValue === CodeNameCodeSequenceValues.Procedure
-  );
-  if (!Procedure) {
-    // console.log('procedure report does not exist leena');
 
-    return [];
-  }
-  // console.log('procedure report does exist leena');
-  // console.log('leena', Procedure);
-  return [];
-}
 function _getMeasurements(ImagingMeasurementReportContentSequence) {
   console.log(ImagingMeasurementReportContentSequence);
   const ImagingMeasurements = ImagingMeasurementReportContentSequence.find(
